+++ src/orders/views/OrderReturn/OrderReturn.tsx
@@ -11,28 +11,10 @@ import OrderReturnPage from "@saleor/orders/components/OrderReturnPage";
 import { OrderReturnFormData } from "@saleor/orders/components/OrderReturnPage/form";
 import { orderUrl } from "@saleor/orders/urls";
 import React from "react";
-import { defineMessages, useIntl } from "react-intl";
+import { useIntl } from "react-intl";
 
+import { messages } from "./messages";
 import ReturnFormDataParser from "./utils";
-
-export const messages = defineMessages({
-  cannotRefundDescription: {
-    id: "XQBVEJ",
-    defaultMessage:
-      "We’ve encountered a problem while refunding the products. Product’s were not refunded. Please try again.",
-    description: "order return error description when cannot refund",
-  },
-  cannotRefundTitle: {
-    id: "l9Lwjh",
-    defaultMessage: "Couldn't refund products",
-    description: "order return error title when cannot refund",
-  },
-  successAlert: {
-    id: "/z9uo1",
-    defaultMessage: "Successfully returned products!",
-    description: "order returned success message",
-  },
-});
 
 interface OrderReturnProps {
   orderId: string;
